try to fix smart_ptr crash
When blocks are requested sometimes the merged-mined part of the block
(auxpow) is invalid and causes a crash in boost dereference for
shared_ptr (variable auxpow) stored in the block and cdiskindex for
writing blocks to disk. I check for null here to see if I can stop it
from crashing because the android wallet can't receive any blocks
otherwise.

Also add NDEBUG define to make release build for qt and daemon.

Updated readme how to build for win32.